pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages=["ragdaemon"]

[project]
name = "ragdaemon"
version = "0.8.0"
description = "Generate and render a call graph for a Python project."
readme = "README.md"
dependencies = [
    "astroid==3.2.2",
    "chromadb==0.4.24",
    "dict2xml==1.7.5",
    "docker==7.1.0",
    "fastapi==0.109.2",
    "Jinja2==3.1.3",
    "networkx==3.2.1",
    "psycopg2-binary==2.9.9",
    "rank_bm25==0.2.2",
    "sqlalchemy==2.0.30",
    "spiceai~=0.3.0",
    "starlette==0.36.3",
    "tqdm==4.66.2",
    "uvicorn==0.29.0",
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/AbanteAI/ragdaemon"

[project.scripts]
ragdaemon = "ragdaemon.__main__:run"

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright==1.1.372",
    "pytest",
    "pytest-asyncio"
]

[tool.pyright]
ignore = ["tests/sample", "venv", ".venv"]
